requirements.txt
dash==2.18.1
dash-bootstrap-components==1.6.0
plotly==5.24.1
paho-mqtt==2.1.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9